% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{flt_connection}
\alias{flt_connection}
\title{Make connection object}
\usage{
flt_connection(host = "127.0.0.1", port = 443, apitoken)
}
\arguments{
\item{host}{Fleet host}

\item{port}{Fleet port (default 443)}

\item{apitoken}{Fleet instance API token}
}
\value{
con object
}
\description{
Make connection object
}
