% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_delete}
\alias{query_delete}
\title{Delete Fleet query by id}
\usage{
query_delete(flt_con, id)
}
\arguments{
\item{flt_con}{connection object via flt_connection()}

\item{id}{Fleet query id, can be fetched via query_list() function}
}
\value{
TRUE if success, else FALSE
}
\description{
Delete Fleet query by id
}
