% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_list}
\alias{query_list}
\title{List queries}
\usage{
query_list(flt_con, req_params = list())
}
\arguments{
\item{flt_con}{connection object via flt_connection()}

\item{req_params}{opt, if provided specifies extra params, i.e. req_params = list(query = "darwin", order_direction = "desc")}
}
\value{
df
}
\description{
List queries
}
