% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_list}
\alias{query_list}
\title{List queries}
\usage{
query_list(flt_con, ...)
}
\arguments{
\item{flt_con}{connection object via flt_connection()}

\item{...}{}
}
\value{
df
}
\description{
List queries
}
